requirements.txt
playwright
pytest
allure-pytest
pytest-playwright
pytest-ordering
pytest-xdist
jsonpath
loguru
python-slugify
allure-python-commons
requests
PyYAML